+++ bingads/manifest.py
@@ -1,7 +1,5 @@
 import sys
-VERSION = '11.5.7'
-BULK_FORMAT_VERSION = '3.0'
-BULK_FORMAT_VERSION_4 = '4.0'
+VERSION = '11.5.8'
 BULK_FORMAT_VERSION_5 = '5.0'
 WORKING_NAME = 'BingAdsSDKPython'
 USER_AGENT = '{0} {1} {2}'.format(WORKING_NAME, VERSION, sys.version_info[0:3])