+++ bingads/v11/bulk/entities/audiences/bulk_in_market_audience.py
@@ -65,6 +65,11 @@ class BulkInMarketAudience(_SingleRecordBulkEntity):
             field_to_csv=lambda c: bulk_str(c.in_market_audience.Scope),
             csv_to_field=lambda c, v: setattr(c.in_market_audience, 'Scope', v if v else None)
         ),
+        _SimpleBulkMapping(
+            _StringTable.AudienceSearchSize,
+            field_to_csv=lambda c: bulk_str(c.in_market_audience.SearchSize),
+            csv_to_field=lambda c, v: setattr(c.in_market_audience, 'SearchSize', int(v) if v else None)
+        ),
     ]
 
     @property